Move: EyeRingCollisionState 파일 분리
