+++ APPRO/APPRO/Views/Stretchings/Eye/Model/EyeStretchingRingEntity.swift
@@ -141,28 +141,3 @@ extension EyeStretchingRingEntity: HasChildren {
     }
     
 }
-
-struct EyeRingCollisionState {
-    
-    let restrictLineCollided: Bool
-    let innerPlaneCollided: Bool
-    
-    var eyesAreInside: Bool {
-        innerPlaneCollided && !restrictLineCollided
-    }
-    
-    func replacing(restrictLine: Bool) -> Self {
-        .init(
-            restrictLineCollided: restrictLine,
-            innerPlaneCollided: innerPlaneCollided
-        )
-    }
-    
-    func replacing(innerPlane: Bool) -> Self {
-        .init(
-            restrictLineCollided: restrictLineCollided,
-            innerPlaneCollided: innerPlane
-        )
-    }
-    
-}
